Plot the delay between two consecutive measurements
Add a chart to plot the delay between two consecutive measurements.
Permits evaluating whether going to a plugin vs a tab in a browser
is necessary to have enough samples.

A first experiment validates that when going to another tab,
no samples seem to be computed (huge increase in the delay).